+++ LemonLime.CTR/Memory.cs
