+++ LemonLime.sln
@@ -7,6 +7,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LemonLime", "LemonLime\Lemo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LemonLime.ARM", "LemonLime.ARM\LemonLime.ARM.csproj", "{65067963-DDB5-4429-9DBE-CCAFBB3FAF56}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LemonLime.CTR", "LemonLime.CTR\LemonLime.CTR.csproj", "{FC9028B7-8D05-4C98-906D-71273FC86979}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LemonLime.Common", "LemonLime.Common\LemonLime.Common.csproj", "{7E6C8E19-7EEF-4901-87A7-553F5B2B56C3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +25,14 @@ Global
 		{65067963-DDB5-4429-9DBE-CCAFBB3FAF56}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{65067963-DDB5-4429-9DBE-CCAFBB3FAF56}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{65067963-DDB5-4429-9DBE-CCAFBB3FAF56}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FC9028B7-8D05-4C98-906D-71273FC86979}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FC9028B7-8D05-4C98-906D-71273FC86979}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FC9028B7-8D05-4C98-906D-71273FC86979}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FC9028B7-8D05-4C98-906D-71273FC86979}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7E6C8E19-7EEF-4901-87A7-553F5B2B56C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7E6C8E19-7EEF-4901-87A7-553F5B2B56C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7E6C8E19-7EEF-4901-87A7-553F5B2B56C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7E6C8E19-7EEF-4901-87A7-553F5B2B56C3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
