+++ LemonLime/LemonLime.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\LemonLime.ARM\LemonLime.ARM.csproj" />
+    <ProjectReference Include="..\LemonLime.Common\LemonLime.Common.csproj" />
   </ItemGroup>
 
 </Project>
