--- LemonLime/Settings.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace LemonLime
-{
-    class Settings
-    {
-        public const int FCRAMSize = 128; // FCRAM size (in MB)
-    }
-}
